+++ src/components/project-cooperation/index.ts
@@ -1,0 +1,1 @@
+export * from './project-cooperation';
