+++ src/components/project-cooperation/project-cooperation.module.css
@@ -1,4 +1,4 @@
-.project-page-cooperation {
+.projectCooperation {
   display: flex;
   flex-direction: column;
 }
