+++ src/components/project-cooperation/project-cooperation.stories.module.css
@@ -1,4 +1,4 @@
-.stories-container {
+.storiesСontainer {
   width: scale(553px);
   margin: scale(92px auto 169px 665px);
