+++ src/components/project-cooperation/project-cooperation.tsx
@@ -1,20 +1,20 @@
 import {FC, HTMLAttributes} from 'react';
 import classNames from 'classnames/bind';
-import styles from './project-page-cooperation.module.css';
+import styles from './project-cooperation.module.css';
 import { Email } from 'shared/types/common';
 import { Icon } from '../ui/icon/icon';
 
 const cx = classNames.bind(styles);
 
-interface IProjectPageCooperationProps extends HTMLAttributes<HTMLElement> {
+interface IProjectCooperationProps extends HTMLAttributes<HTMLElement> {
   email: Email;
 }
 
-export const ProjectPageCooperation: FC<IProjectPageCooperationProps> = (props) => {
+export const ProjectCooperation: FC<IProjectCooperationProps> = (props) => {
   const { email } = props;
 
   return (
-    <div className={cx('project-page-cooperation')}>
+    <div className={cx('projectCooperation')}>
       <Icon glyph='asterisk' className={cx('asterisk')} />
       <h3 className={cx('header')}>Проект открыт к&nbsp;сотрудничеству</h3>
       <p className={cx('text')}>Мы находимся в постоянном поиске режиссёров и актеров, заинтересованных в постановке читок.</p>
