--- src/components/project-page-cooperation/index.ts
@@ -1,1 +1,0 @@
-export * from './project-page-cooperation';
